+++ src/main/java/com/capston/favicon/user/application/RedisServiceImpl.java
@@ -1,6 +1,6 @@
-package com.capston.favicon.application;
+package com.capston.favicon.user.application;
 
-import com.capston.favicon.application.repository.RedisService;
+import com.capston.favicon.user.application.service.RedisService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
