+++ src/main/java/com/capston/favicon/user/application/service/MailService.java
@@ -1,4 +1,4 @@
-package com.capston.favicon.application.repository;
+package com.capston.favicon.user.application.service;
 
 public interface MailService {
     void send(String to, String subject, String text);
