+++ src/main/java/com/capston/favicon/user/controller/DataController.java
@@ -1,10 +1,10 @@
-package com.capston.favicon.presentation.controller;
+package com.capston.favicon.user.controller;
 
 
-import com.capston.favicon.application.repository.DataService;
+import com.capston.favicon.user.application.service.DataService;
 import com.capston.favicon.config.APIResponse;
-import com.capston.favicon.domain.domain.Data;
-import com.capston.favicon.domain.dto.SearchDto;
+import com.capston.favicon.user.domain.Data;
+import com.capston.favicon.user.dto.SearchDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
