+++ src/main/java/com/capston/favicon/user/domain/Data.java
@@ -1,4 +1,4 @@
-package com.capston.favicon.domain.domain;
+package com.capston.favicon.user.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
